% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tsk_CvsR}
\alias{tsk_CvsR}
\title{Creates Circle vs Ring Problem}
\usage{
tsk_CvsR(N = 100, nPoints = 100, r0 = 1, dr = 0.1, sdDev = 0.1)
}
\arguments{
\item{N}{number of point clouds}

\item{nPoints}{number of points in each point cloud}

\item{r0}{radius of the enter circle}

\item{dr}{half of the distance between two circles in the ring}

\item{sdDev}{measures the closeness of points to original shapes}
}
\description{
This ffunction MLR3 task with circle vs ring classification problem
}
\details{
Each data entry is a cloud of point scattered around either a single circle with radius r0
or a ring (i.e. two circles with radia r0+dr and r0-dr)
The problem would be to guess type of the point cloud
}
