% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tsk_ellipse}
\alias{tsk_ellipse}
\title{Creates ellepse regression task}
\usage{
tsk_ellipse(N = 100, sdDev = 0.1)
}
\arguments{
\item{N}{number of points}

\item{sdDev}{specifies how far are points from the ellipse}
}
\value{
MLR3 task object with r (list of squzzed ratios) and X (list of point clouds) data fields
}
\description{
This function creates MLR3 task with squzzed ellipse regression problem
}
